kernel: Reintegrate terminal and add integration test
