--- test/util/integration/term/Makefile
@@ -1,19 +1,0 @@
-#################################################
-#          IncludeOS SERVICE makefile           #
-#################################################
-
-# The name of your service
-SERVICE = test_terminal
-SERVICE_NAME = The Terminal Test Service
-
-# Your service parts
-FILES=term.cpp
-# Your disk image
-DISK=
-
-# IncludeOS location
-ifndef INCLUDEOS_INSTALL
-INCLUDEOS_INSTALL=$(HOME)/IncludeOS_install
-endif
-
-include $(INCLUDEOS_INSTALL)/Makeseed
