--- test/util/integration/term/banana.txt
@@ -1,13 +1,0 @@
-     ____                           ___
-    |  _ \  ___              _   _.' _ `.
- _  | [_) )' _ `._   _  ___ ! \ | | (_) |    _
-|:;.|  _ <| (_) | \ | |' _ `|  \| |  _  |  .:;|
-|   `.[_) )  _  |  \| | (_) |     | | | |.',..|
-':.   `. /| | | |     |  _  | |\  | | |.' :;::'
- !::,   `-!_| | | |\  | | | | | \ !_!.'   ':;!
- !::;       ":;:!.!.\_!_!_!.!-'-':;:''    '''!
- ';:'        `::;::;'             ''     .,  .
-   `:     .,.    `'    .::... .      .::;::;'
-     `..:;::;:..      ::;::;:;:;,    :;::;'
-       "-:;::;:;:      ':;::;:''     ;.-'
-           ""`---...________...---'""
